Move argument parsing logic into a lib file
