+++ scripts/metadata_comparison/metadata_comparison/extractor.py
@@ -25,6 +25,7 @@ from google.cloud import storage
 import google.auth
 from googleapiclient.discovery import build as google_client_build
 import logging as log
+from metadata_comparison.lib_argument_regex import *
 
 logger = log.getLogger('metadata_comparison.extractor')
 
@@ -38,67 +39,6 @@ def set_log_verbosity(verbose):
 def quieten_chatty_imports():
     log.getLogger('googleapiclient.discovery_cache').setLevel(log.ERROR)
     log.getLogger('googleapiclient.discovery').setLevel(log.WARNING)
-
-
-def workflow_regex_validator(value):
-    """Makes sure that a value is a valid Cromwell workflow ID then returns the workflow ID"""
-    workflow_regex=re.compile('^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$')
-    if not workflow_regex.match(value):
-        msg = f'Invalid workflow ID {value}. Expected {workflow_regex.pattern}'
-        raise argparse.ArgumentTypeError(msg)
-    else:
-        return value
-
-
-def url_regex_validator(value):
-    """
-    Validates then extract the root of the Cromwell URL from the various URL strings which might be provided.
-    Deliberately flexible because it's tedious to remember which script requires which type of format.
-    eg:
-        'http://localhost' => 'http://localhost'
-        'http://localhost:8000' => 'http://localhost:8000'
-        'http://localhost:8000/' => 'http://localhost:8000'
-        'http://localhost:8000/api/workflows/' => 'http://localhost:8000'
-        'http://localhost:8000/custom/prefix/api/workflows/' => 'http://localhost:8000/custom/prefix'
-    """
-    url_regex = re.compile('(http(s?)://((?!/api).)*[^/])(/(api.*)?)?')
-    m = url_regex.match(value)
-    if m:
-        return m.group(1)
-    else:
-        msg = f'Invalid Cromwell URL {value}. Expected {url_regex.pattern}'
-        raise argparse.ArgumentTypeError(msg)
-
-
-def gcs_path_regex_validator(value):
-    """
-    Validates then extracts the bucket and object-path from a GS string. Returned as a pair.
-    eg:
-        'gs://bucket/path/to/directory/' -> ('bucket', 'path/to/directory')
-    """
-    gcs_regex = re.compile('^gs://([a-zA-Z0-9-]+)/(([a-zA-Z0-9-]+/)*[a-zA-Z0-9-]+)/?$')
-    m = gcs_regex.match(value)
-    if m:
-        return m.group(1), m.group(2)
-    else:
-        msg = f'Invalid GCS path {value}. Expected {gcs_regex.pattern}'
-        raise argparse.ArgumentTypeError(msg)
-
-
-def get_operation_id_number(value):
-    """
-    Validates then extracts from PAPI operation IDs just the final number.
-    eg:
-        'projects/project_name/operations/01234567891011121314' -> '01234567891011121314'
-    """
-
-    operation_regex = re.compile('^.*/([^/]*)')
-    m = operation_regex.search(value)
-    if m:
-        return m.group(1)
-    else:
-        msg = f'Unexpected operation ID {value}. Expected something like {operation_regex.pattern}'
-        raise Exception(msg)
 
 
 def uploadLocalCheckout():
